1.RSF开发完成，进入测试阶段。

+++ hasor-rsf/src/main/java/net/hasor/rsf/protocol/toos/ProtocolUtils.java
@@ -1,3 +1,18 @@
+/*
+ * Copyright 2008-2009 the original 赵永春(zyc@hasor.net).
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package net.hasor.rsf.protocol.toos;
 import java.lang.reflect.Array;
 import net.hasor.rsf.general.ProtocolStatus;
@@ -37,7 +52,7 @@ public class ProtocolUtils {
     //
     /**判断 response 是否为一个ACK包。*/
     public static boolean isACK(ResponseSocketBlock socketMessage) {
-        return socketMessage.getStatus() == ProtocolStatus.Accepted.shortValue();
+        return socketMessage.getStatus() == ProtocolStatus.Accepted;
     }
     /**是否为Request消息。*/
     public static boolean isRequest(byte version) {
